# lib/api-zod/src/generated/types/appHealth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppHealth {
  status: string;
  hasKey: boolean;
  source?: string;
  cacheKeys?: number;
  uptime?: number;
}