# lib/api-zod/src/generated/types/signalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignalRequest {
  ticker: string;
  asset: string;
  buys: number;
  sells: number;
  totalValue: number;
  members: string[];
  flags: string[];
  committees: string[];
}