# lib/api-zod/src/generated/types/signalResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignalResult {
  signal: string;
  confidence: number;
  summary: string;
  flag_note: string;
}